backend/requirements.txt
asgiref==3.2.10
Django==3.1.1
django-environ==0.4.5
djangorestframework==3.11.1
psycopg2==2.8.6
pytz==2020.1
sqlparse==0.3.1
